Remove prometheus ServiceMonitor from k8s manifest
For some reason, the team thought it's a great idea to change the default...
This leads to peoples deployments FAILING when they do not have prometheus in their cluster.

Who the fuck was like "oh yeah let's add this feature... and turn it on by default :)"
No prometheus => failed deployment. It literally breaks as easily as that.